feat(core): support tier environment and fractional cpu

+++ internal/api/admin_handlers.go
@@ -249,7 +249,7 @@ func (h *APIHandlers) HandleAdminAppTiers(w http.ResponseWriter, r *http.Request
 	case http.MethodPost, http.MethodPut:
 		var req struct {
 			Name         string                `json:"name"`
-			CPU          int                   `json:"cpu"`
+			CPU          float64               `json:"cpu"`
 			Memory       string                `json:"memory"`
 			Storage      string                `json:"storage"`
 			PriceMonthly float64               `json:"price_monthly"`
